+++ hammer-sqldb/src/main/java/cn/featherfly/hammer/sqldb/jdbc/JdbcProcedureMulitiQuery.java
@@ -8,17 +8,17 @@
  */
 package cn.featherfly.hammer.sqldb.jdbc;
 
-import java.io.IOException;
 import java.io.Serializable;
-import java.sql.CallableStatement;
 import java.sql.ResultSet;
 import java.sql.SQLException;
+import java.sql.Statement;
 import java.util.List;
 import java.util.Map;
 import java.util.NoSuchElementException;
 import java.util.function.Function;
 
 import cn.featherfly.common.db.JdbcException;
+import cn.featherfly.common.db.JdbcUtils;
 import cn.featherfly.common.db.mapper.SqlResultSet;
 import cn.featherfly.common.db.mapping.SqlTypeMappingManager;
 import cn.featherfly.common.repository.MulitiQuery;
@@ -29,7 +29,7 @@ import cn.featherfly.common.repository.mapper.RowMapper;
  *
  * @author zhongj
  */
-public class JdbcProcedureMulitiQuery implements MulitiQuery, AutoCloseable {
+public class JdbcProcedureMulitiQuery implements MulitiQuery {
 
     private int index = 0;
 
@@ -39,7 +39,7 @@ public class JdbcProcedureMulitiQuery implements MulitiQuery, AutoCloseable {
 
     private SqlTypeMappingManager manager;
 
-    private CallableStatement call;
+    private Statement stat;
 
     private Function<Class<?>, RowMapper<?>> getTypeMapper;
 
@@ -48,15 +48,15 @@ public class JdbcProcedureMulitiQuery implements MulitiQuery, AutoCloseable {
     /**
      * Instantiates a new jdbc muliti query.
      *
-     * @param call the call
+     * @param stat the call
      * @param manager the manager
      * @param getTypeMapper the get type mapper
      * @param postHandler the post handler
      */
-    public JdbcProcedureMulitiQuery(CallableStatement call, SqlTypeMappingManager manager,
+    public JdbcProcedureMulitiQuery(Statement stat, SqlTypeMappingManager manager,
         Function<Class<?>, RowMapper<?>> getTypeMapper, Function<List<?>, ?> postHandler) {
         super();
-        this.call = call;
+        this.stat = stat;
         this.manager = manager;
         this.getTypeMapper = getTypeMapper;
         this.postHandler = postHandler;
@@ -86,7 +86,9 @@ public class JdbcProcedureMulitiQuery implements MulitiQuery, AutoCloseable {
     @Override
     public <E> List<E> next(RowMapper<E> mapper) {
         RowMapperResultSetExtractor<E> extractor = new RowMapperResultSetExtractor<>(mapper);
-        List<E> list = extractor.extract(new SqlResultSet(nextResultSet()));
+        ResultSet res = nextResultSet();
+        List<E> list = extractor.extract(new SqlResultSet(res));
+        JdbcUtils.close(res); // stat可能不是CascadedCloseStatement，所以先手动关闭 
         return (List<E>) postHandler.apply(list);
     }
 
@@ -94,7 +96,7 @@ public class JdbcProcedureMulitiQuery implements MulitiQuery, AutoCloseable {
         try {
             if (index == 0) {
                 index++;
-                return call.getResultSet();
+                return stat.getResultSet();
             }
 
             if (!forward) { // 表示没有调用hasNext，即没有调用call.getMoreResults()
@@ -108,7 +110,7 @@ public class JdbcProcedureMulitiQuery implements MulitiQuery, AutoCloseable {
             forward = false;
             index++;
             // cursor已经移动到位置了，可以直接取值
-            return call.getResultSet();
+            return stat.getResultSet();
         } catch (SQLException e) {
             throw new JdbcException(e);
         }
@@ -134,7 +136,7 @@ public class JdbcProcedureMulitiQuery implements MulitiQuery, AutoCloseable {
             return hasNext;
         }
         try {
-            hasNext = call.getMoreResults();
+            hasNext = stat.getMoreResults();
             forward = true;
             return hasNext;
         } catch (SQLException e) {
@@ -146,11 +148,10 @@ public class JdbcProcedureMulitiQuery implements MulitiQuery, AutoCloseable {
      * {@inheritDoc}
      */
     @Override
-    public void close() throws IOException {
+    public void close() throws JdbcException {
         try {
-            if (!call.isClosed() && !call.getConnection().isClosed()) {
-                // AutoCloseConnection, auto close it create object
-                call.getConnection().close();
+            if (!stat.isClosed()) {
+                stat.close();
             }
         } catch (SQLException e) {
             throw new JdbcException(e);