+++ hammer-sqldb/src/test/java/cn/featherfly/hammer/sqldb/pt/BenchmarkTestBase.java
@@ -8,7 +8,7 @@ import java.sql.SQLException;
 
 import javax.sql.DataSource;
 
-import org.apache.commons.dbcp.BasicDataSource;
+import org.apache.commons.dbcp2.BasicDataSource;
 import org.apache.logging.log4j.core.config.Configurator;
 import org.testng.annotations.BeforeSuite;
 import org.testng.annotations.Optional;